moved the globe into its own directory
